fixed (kind of) the functionality of the freeze effect
Fixed (kind of) the freeze functionality so that different objects with different components can be frozen by the freeze effect regardless of which objects they have. Still need to work on the colliders attached to the ice containers so that they dont fly off. should also fix the code and make it more neat.